Fix cursor positioning issue when moving to the end of a full line.
If the line was completely full and the cursor was not at the end of the
line (e.g. inserting from the beginning of the line), and then the
cursor was moved to the end of the line, the cursor should have gone
onto the next line. This wasn't happening on the last line. Instead, the
cursor would not create the new line and the cursor would wind up being
located one line up the screen from where it should have been.
Fixed this by adding a check into minirl_refresh_cursor() to see if the
cursor would move onto a brand new line, and if so then a full refresh
is performed, which will add lines as necessary.