Make a new line when line filled
If the last line of a multi-line line is filled and the cursor is not on
that line, the next prompt wasn't being positioned after the last line;
it was being positioned on the line after the cursor position.
This change gets a new line created as soon as a line is filled even if
the edit position isn't at the end of the line.
It still doesn't work quite right as the next prompt will be placed on
the last line, not after the last line, but this is still much better
than the previous behaviour. Not tested that well, which is why it
hasn't been merged into master.