Update UTF-8 input files
Updated the UTF-8 input files and updated the associated utf-8 code as
necessary to get the handling of new character types to work as desired.
Specifically, the character u200D (zero width joiner) character wasn't
being handled correctly (was being give a length of 1), and wasn't being
associated with the adjopining character to its left, which is what
bash on linux seems to do. The result was that an invisible character
existed in the input line, which could be deleted separately from any
character to its left.
Also made a change so that when deleting a character to the left of the
current edit position the grapheme to the left is removed, not the
character to the left (which is what tinyrl does). This makes things
consistent with deleting characters to the right of the edit position.