minirl: Improve cursor column positioning
When refreshing the line there is no need to move the cursor after
printing the line unless the edit point isn't at the end of the line.
Rather than unconditionally moving the cursor to the desired row/column,
only move it if it needs to be moved. In the common case where the line
is refreshed when initially displaying the prompt, it will mean the
cursor isn't moved at all, avoiding outputting unnecessary escape
sequences.