emit '\n' when line edit completes
Rather than expect the user to emit the '\n' manually when the enter
key is pressed (as is the case with DAL CLI), minirl will emit the '\n'
after the cursor has been positioned at the end of the edit line.
The problem was that the edit position might not have been at the end of
the edit line when enter was pressed, so emitting a '\n' might not have
given the expected result.